+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed "calling remove() for dragging shape will throw an error"
 - Fixed wrong opacity level for cached group with opacity
 - More consistent shadows on HDPI screens
+- Fixed memory leak for nodes with several names
 
 ## [1.2.2][2016-09-15]